Move all relative timestamp display text generation to a single place
